move to _common
